services/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.12.0
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   openai
    #   starlette
ariadne[file-uploads]==0.20.1
    # via -r requirements.in
arrow==1.3.0
    # via -r requirements.in
assemblyai==0.20.2
    # via -r requirements.in
base58==2.1.1
    # via
    #   scalecodec
    #   substrate-interface
boto3==1.28.64
    # via -r requirements.in
botocore==1.31.85
    # via
    #   boto3
    #   s3transfer
cachecontrol==0.13.1
    # via firebase-admin
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
    #   substrate-interface
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
cryptography==41.0.7
    # via pyjwt
cytoolz==0.12.3
    # via eth-utils
distro==1.9.0
    # via openai
ecdsa==0.18.0
    # via
    #   -r requirements.in
    #   substrate-interface
eth-hash==0.7.0
    # via eth-utils
eth-keys==0.5.1
    # via substrate-interface
eth-typing==5.0.0
    # via
    #   eth-keys
    #   eth-utils
eth-utils==2.3.1
    # via
    #   eth-keys
    #   substrate-interface
fastapi==0.103.2
    # via -r requirements.in
firebase-admin==6.2.0
    # via -r requirements.in
google-api-core[grpc]==2.14.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.108.0
    # via firebase-admin
google-auth==2.23.4
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.1
    # via google-api-python-client
google-cloud-core==2.3.3
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.13.1
    # via firebase-admin
google-cloud-storage==2.13.0
    # via firebase-admin
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.6.0
    # via google-cloud-storage
googleapis-common-protos==1.61.0
    # via
    #   google-api-core
    #   grpcio-status
graphql-core==3.2.3
    # via ariadne
grpcio==1.59.3
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.59.3
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.26.0
    # via
    #   assemblyai
    #   openai
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   substrate-interface
invoke==2.2.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mako==1.2.4
    # via alembic
mangum==0.17.0
    # via -r requirements.in
markupsafe==2.1.3
    # via mako
more-itertools==10.4.0
    # via scalecodec
msgpack==1.0.7
    # via cachecontrol
mutagen==1.47.0
    # via -r requirements.in
openai==1.12.0
    # via -r requirements.in
pillow==10.1.0
    # via -r requirements.in
proto-plus==1.22.3
    # via google-cloud-firestore
protobuf==4.25.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
py-bip39-bindings==0.1.11
    # via substrate-interface
py-ed25519-zebra-bindings==1.0.1
    # via substrate-interface
py-sr25519-bindings==0.2.0
    # via substrate-interface
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pycryptodome==3.20.0
    # via
    #   -r requirements.in
    #   substrate-interface
pydantic==2.3.0
    # via
    #   -r requirements.in
    #   assemblyai
    #   fastapi
    #   openai
pydantic-core==2.6.3
    # via pydantic
pyjwt[crypto]==2.8.0
    # via firebase-admin
pynacl==1.5.0
    # via substrate-interface
pyparsing==3.1.1
    # via httplib2
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
    #   scalecodec
    #   substrate-interface
rsa==4.9
    # via google-auth
s3transfer==0.7.0
    # via boto3
scalecodec==1.2.11
    # via substrate-interface
schedule==1.2.1
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.22
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via -r requirements.in
starlette==0.27.0
    # via
    #   ariadne
    #   fastapi
substrate-interface==1.7.10
    # via -r requirements.in
toolz==0.12.1
    # via cytoolz
tqdm==4.66.2
    # via openai
types-python-dateutil==2.8.19.14
    # via arrow
typing-extensions==4.8.0
    # via
    #   alembic
    #   ariadne
    #   assemblyai
    #   eth-typing
    #   fastapi
    #   mangum
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.0.7
    # via
    #   botocore
    #   requests
uvicorn==0.23.2
    # via -r requirements.in
websocket-client==1.8.0
    # via substrate-interface
websockets==12.0
    # via assemblyai
xxhash==3.5.0
    # via substrate-interface
